Added Error Handling for User Login
Added error handling for when the user unsuccessfully logs in.  It checks for the expected condition (ID "duo form") - if this is present it breaks out of the loop and continues execution.  If the expected condition is not present, it prints a message and clears out the "uniqname" input field and prepares to resend another user login attempt.